# docs/api-spec.yml
openapi: 3.0.3
info:
  title: "Líder Digital API"
  description: |
    API para gestión de usuarios, blog y captación de leads.
    Autenticación mediante Firebase Auth.
    Los mensajes de respuesta están en español latinoamericano neutro.
  version: "1.0.0"

servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
  - url: https://api.liderdigital.cl/v1
    description: Producción

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "El correo electrónico no es válido"
        statusCode:
          type: integer
          example: 400

    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"
        timestamp:
          type: string
          format: date-time

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Verificación de salud de la API
      security: []
      tags:
        - Sistema
      responses:
        "200":
          description: API operativa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /auth/login:
    post:
      summary: Iniciar sesión con Firebase ID token
      security: []
      tags:
        - Autenticación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_token
              properties:
                id_token:
                  type: string
                  description: Firebase ID token del cliente
      responses:
        "200":
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
                      token:
                        type: string
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Inicio de sesión exitoso"
        "401":
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/me:
    get:
      summary: Obtener usuario autenticado actual
      tags:
        - Autenticación
      responses:
        "200":
          description: Usuario obtenido exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users:
    get:
      summary: Listar usuarios (Solo admin)
      tags:
        - Usuarios
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  meta:
                    $ref: "#/components/schemas/Pagination"
        "403":
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Crear usuario (Solo admin)
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - full_name
                - role
              properties:
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                role:
                  type: string
                  enum: [admin, editor]
      responses:
        "201":
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Usuario creado exitosamente"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users/{id}:
    get:
      summary: Obtener usuario por ID (Solo admin)
      tags:
        - Usuarios
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usuario obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Actualizar usuario (Solo admin)
      tags:
        - Usuarios
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                role:
                  type: string
                  enum: [admin, editor]
                is_active:
                  type: boolean
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Usuario actualizado exitosamente"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /categories:
    get:
      summary: Listar categorías
      security: []
      tags:
        - Categorías
      responses:
        "200":
          description: Lista de categorías
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
        "500":
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Crear categoría (Admin/Editor)
      tags:
        - Categorías
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                slug:
                  type: string
                  pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
                description:
                  type: string
      responses:
        "201":
          description: Categoría creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Categoría creada exitosamente"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /categories/{id}:
    patch:
      summary: Actualizar categoría (Admin/Editor)
      tags:
        - Categorías
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Categoría actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Categoría actualizada exitosamente"
        "404":
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Eliminar categoría (Solo admin)
      tags:
        - Categorías
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Categoría eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Categoría eliminada exitosamente"
        "404":
          description: Categoría no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /articles:
    get:
      summary: Listar artículos publicados
      security: []
      tags:
        - Artículos
      parameters:
        - name: category_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de artículos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Article"
                  meta:
                    $ref: "#/components/schemas/Pagination"

    post:
      summary: Crear artículo (Admin/Editor)
      tags:
        - Artículos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - slug
                - content
                - cover_url
                - category_id
                - status
              properties:
                title:
                  type: string
                  minLength: 5
                  maxLength: 150
                slug:
                  type: string
                content:
                  type: string
                cover_url:
                  type: string
                  format: uri
                category_id:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [draft, published]
      responses:
        "201":
          description: Artículo creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Article"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Artículo creado exitosamente"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /articles/{id}:
    get:
      summary: Obtener artículo por ID o slug
      security: []
      tags:
        - Artículos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artículo obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Article"
        "404":
          description: Artículo no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Actualizar artículo (Admin/Editor)
      tags:
        - Artículos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                slug:
                  type: string
                content:
                  type: string
                cover_url:
                  type: string
                category_id:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [draft, published]
      responses:
        "200":
          description: Artículo actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Article"
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Artículo actualizado exitosamente"
        "404":
          description: Artículo no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Eliminar artículo (Solo admin)
      tags:
        - Artículos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artículo eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Artículo eliminado exitosamente"
        "404":
          description: Artículo no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /leads/subscribe:
    post:
      summary: Registrar suscriptor al newsletter
      security: []
      tags:
        - Leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - full_name
              properties:
                email:
                  type: string
                  format: email
                  description: Email del suscriptor
                full_name:
                  type: string
                  minLength: 1
                  description: Nombre completo del suscriptor
      responses:
        "201":
          description: Suscriptor registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "¡Gracias por suscribirte!"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /leads/contact:
    post:
      summary: Registrar mensaje de contacto
      security: []
      tags:
        - Leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - full_name
                - message
              properties:
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                  minLength: 1
                company:
                  type: string
                  description: Empresa (opcional)
                message:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Mensaje registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      lead_id:
                        type: string
                        format: uuid
                  meta:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Tu mensaje ha sido enviado. Nos pondremos en contacto contigo pronto."
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /leads:
    get:
      summary: Listar leads (Solo admin)
      tags:
        - Leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [subscriber, contact]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  meta:
                    $ref: "#/components/schemas/Pagination"
        "403":
          description: Acceso denegado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: Firebase UID
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, editor]
        avatar_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        cover_url:
          type: string
          format: uri
        category_id:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        author_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        status:
          type: string
          enum: [subscriber, contact]
        metadata:
          type: object
        attribution:
          type: object
        payload:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time